feat: :sparkles: Create tab view for main profile update and tags preference update.
